triton kernel offset to(tl.int64)

+++ src/paddlefleet/fusions/fused_mhc_kernels.py
@@ -106,7 +106,7 @@ if _TRITON_AVAILABLE:
         NUM_ITERS: tl.constexpr,
     ):
         """Grid: (N_batch,). Each program handles one [HC, HC] matrix."""
-        pid = tl.program_id(0)
+        pid = tl.program_id(0).to(tl.int64)
         if pid >= N_batch:
             return
         base = pid * HC * HC
@@ -142,7 +142,7 @@ if _TRITON_AVAILABLE:
         NUM_ITERS: tl.constexpr,
     ):
         """Grid: (N_batch,). Each program handles one [HC, HC] backward."""
-        pid = tl.program_id(0)
+        pid = tl.program_id(0).to(tl.int64)
         if pid >= N_batch:
             return
         base = pid * HC * HC
@@ -308,8 +308,8 @@ if _TRITON_AVAILABLE:
         BLOCK_S: tl.constexpr,
     ):
         """out[s, c] = sum_i x[s, i, c] * h[s, i]."""
-        pid_s = tl.program_id(0)
-        pid_c = tl.program_id(1)
+        pid_s = tl.program_id(0).to(tl.int64)
+        pid_c = tl.program_id(1).to(tl.int64)
         offs_s = pid_s * BLOCK_S + tl.arange(0, BLOCK_S)
         offs_c = pid_c * BLOCK_C + tl.arange(0, BLOCK_C)
         mask_s = offs_s < sb
@@ -386,8 +386,8 @@ if _TRITON_AVAILABLE:
         BLOCK_S: tl.constexpr,
     ):
         """out = hr.T @ orig + hp * (x + bias)."""
-        pid_s = tl.program_id(0)
-        pid_c = tl.program_id(1)
+        pid_s = tl.program_id(0).to(tl.int64)
+        pid_c = tl.program_id(1).to(tl.int64)
         offs_s = pid_s * BLOCK_S + tl.arange(0, BLOCK_S)
         offs_c = pid_c * BLOCK_C + tl.arange(0, BLOCK_C)
         mask_s = offs_s < sb
@@ -500,8 +500,8 @@ if _TRITON_AVAILABLE:
         BLOCK_S: tl.constexpr,
     ):
         """g_x = hp @ go, g_orig = hr @ go."""
-        pid_s = tl.program_id(0)
-        pid_c = tl.program_id(1)
+        pid_s = tl.program_id(0).to(tl.int64)
+        pid_c = tl.program_id(1).to(tl.int64)
         offs_s = pid_s * BLOCK_S + tl.arange(0, BLOCK_S)
         offs_c = pid_c * BLOCK_C + tl.arange(0, BLOCK_C)
         mask_s = offs_s < sb
@@ -580,7 +580,7 @@ if _TRITON_AVAILABLE:
         BLOCK_S: tl.constexpr,
     ):
         """g_hp = sum_c go*(x+bias), g_hr = orig @ go.T."""
-        pid_s = tl.program_id(0)
+        pid_s = tl.program_id(0).to(tl.int64)
         offs_s = pid_s * BLOCK_S + tl.arange(0, BLOCK_S)
         mask_s = offs_s < sb
         g_hp_acc = tl.zeros((BLOCK_S, N), dtype=tl.float32)
